matterflow-ai/requirements.txt
fastapi
uvicorn
uvicorn[standard]
python-docx
pdfplumber
pillow
pytesseract
boto3
python-multipart
python-dotenv
transformers 
torch 
pydantic
transformers[torch]
sentencepiece
datasets
accelerate
openai
